simplify the lookup for parent hn
the filter it was using (find_closest_hn_id) was used when *any* mention of an issue,
not just in the heading, was used to the tool.
Now that the selector explicitly looks for anchors *inside* a h1/h2/h3...
there is no point in using find_closest_hn_id, because the hn will always be an ancestor.
The new function find_ancestor_hn_id does not check siblings,
and should therefore be faster.